pyproject.toml
[tool.poetry]
name = "galgoz"
version = "0.1.0"
description = ""
authors = ["Gustavo <gzarruk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.9"
pandas = "^2.2.3"
numpy = "^2.2.0"
flask = "^3.1.0"
oandapyv20 = "^0.7.2"
ipykernel = "^6.29.5"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
pytest = "^8.3.4"
plotly = "^5.24.1"
ipython = "^8.30.0"
nbformat = "^5.10.4"
types-python-dateutil = "^2.9.0.20241206"
pydantic = "^2.10.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
